+++ src/components/FooterInner/FooterInner.module.css
@@ -1,4 +1,4 @@
-.footer {
+.footerInner {
   background-color: var(--footer-bg-color);
 
   .column {
